Revert "Fix for delayed switch drifting (issue #1276)"
