+++ main/Helper.cpp
@@ -39,7 +39,7 @@ void StringSplit(std::string str, const std::string &delim, std::vector<std::str
 
 void stdreplace(
 	std::string &inoutstring,
-	const std::string& replaceWhat,
+	const std::string& replaceWhat, 
 	const std::string& replaceWithWhat)
 {
 	int pos = 0;
@@ -136,7 +136,7 @@ std::vector<std::string> GetSerialPorts(bool &bUseDirectPath)
 				sprintf(szPortName, "COM%d", ii);
 				ret.push_back(szPortName); // add port
 			}
-			// --------------
+			// --------------            
 		}
 	}
 	// Method 3: EnumSerialPortsWindows, often fails
@@ -202,7 +202,7 @@ std::vector<std::string> GetSerialPorts(bool &bUseDirectPath)
 				bUseDirectPath = true;
 				ret.push_back("/dev/" + fname);
 			}
-#ifdef __FreeBSD__
+#ifdef __FreeBSD__            
 			else if (fname.find("ttyU")!=std::string::npos)
 			{
 				bUseDirectPath=true;
@@ -271,7 +271,7 @@ std::vector<std::string> GetSerialPorts(bool &bUseDirectPath)
 
 bool file_exist (const char *filename)
 {
-	struct stat sbuffer;
+	struct stat sbuffer;   
 	return (stat(filename, &sbuffer) == 0);
 }
 
@@ -375,7 +375,7 @@ double CalculateDewPoint(double temp, int humidity)
 	return dew_numer/dew_denom;
 }
 
-uint32_t IPToUInt(const std::string &ip)
+uint32_t IPToUInt(const std::string &ip) 
 {
 	int a, b, c, d;
 	uint32_t addr = 0;
@@ -510,12 +510,12 @@ std::vector<std::string> ExecuteCommandAndReturn(const std::string &szCommand)
 	}
 	catch (...)
 	{
-
+		
 	}
 	return ret;
 }
 
-//convert date string 10/12/2014 10:45:58 en  struct tm
+//convert date string 10/12/2014 10:45:58 en  struct tm 
 void DateAsciiTotmTime (std::string &sTime , struct tm &tmTime  )
 {
 		tmTime.tm_isdst=0; //dayly saving time
@@ -528,7 +528,7 @@ void DateAsciiTotmTime (std::string &sTime , struct tm &tmTime  )
 
 
 }
-//convert struct tm time to char
+//convert struct tm time to char 
 void AsciiTime (struct tm &ltime , char * pTime )
 {
 		sprintf(pTime, "%04d-%02d-%02d %02d:%02d:%02d", ltime.tm_year + 1900, ltime.tm_mon + 1, ltime.tm_mday, ltime.tm_hour, ltime.tm_min, ltime.tm_sec);
@@ -536,7 +536,7 @@ void AsciiTime (struct tm &ltime , char * pTime )
 
 std::string  GetCurrentAsciiTime ()
 {
-	    time_t now = time(0)+1;
+	    time_t now = time(0)+1;	
 		struct tm ltime;
 		localtime_r(&now, &ltime);
 		char pTime[40];
@@ -837,36 +837,3 @@ int gettimeofday( timeval * tp, void * tzp)
 	return 0;
 }
 #endif
-
-int getclock(struct timeval *tv) {
-#ifdef CLOCK_MONOTONIC
-	struct timespec ts;
-		if (!clock_gettime(CLOCK_MONOTONIC, &ts)) {
-			tv->tv_sec = ts.tv_sec;
-			tv->tv_usec = ts.tv_nsec / 1000;
-			return 0;
-		}
-#endif
-	return gettimeofday(tv, NULL);
-}
-int timeval_subtract (struct timeval *result, struct timeval *x, struct timeval *y) {
-	/* Perform the carry for the later subtraction by updating y. */
-  if (x->tv_usec < y->tv_usec) {
-		int nsec = (y->tv_usec - x->tv_usec) / 1000000 + 1;
-    y->tv_usec -= 1000000 * nsec;
-    y->tv_sec += nsec;
-  }
-  if (x->tv_usec - y->tv_usec > 1000000) {
-    int nsec = (x->tv_usec - y->tv_usec) / 1000000;
-    y->tv_usec += 1000000 * nsec;
-    y->tv_sec -= nsec;
-  }
-
-  /* Compute the time remaining to wait.
-     tv_usec is certainly positive. */
-  result->tv_sec = x->tv_sec - y->tv_sec;
-  result->tv_usec = x->tv_usec - y->tv_usec;
-
-  /* Return 1 if result is negative. */
-  return x->tv_sec < y->tv_sec;
-}
