+++ main/Helper.h
@@ -4,7 +4,7 @@
 void StringSplit(std::string str, const std::string &delim, std::vector<std::string> &results);
 void stdreplace(
 	std::string &inoutstring,
-	const std::string& replaceWhat,
+	const std::string& replaceWhat, 
 	const std::string& replaceWithWhat);
 void stdupper(std::string &inoutstring);
 bool file_exist (const char *filename);
@@ -60,5 +60,3 @@ std::string MakeHtml(const std::string &txt);
 #if defined WIN32
 	int gettimeofday(timeval * tp, void * tzp);
 #endif
-int getclock(struct timeval *tv);
-int timeval_subtract (struct timeval *result, struct timeval *x, struct timeval *y);
